refactor:Cloud Pageが長すぎるのを分離
